Phases 1-7: handshake protocol, brand string, timing coordinator, canary, alloc tracker emulation, handshake_test
- Phase 1: Enhanced MagicCpuid handshake protocol (REGISTER_PID, REGISTER_SYSCALL,
  GET_SYSCALL_HANDLER, ENHANCED_MODE, SET_SHM, GET_SHM, per-process PID tracking)
- Phase 2: LSTAR syscall monitoring (syscall handler registration + query via MagicCpuid)
- Phase 3: Brand string CPUID (HandleBrandStringLeaf for leaves 0x80000002-4,
  config-driven brand_string/enhanced_brand_string, enhanced mode override)
- Phase 4: TimingCoordinator (cross-handler RDTSC->CPUID->RDTSC pattern detection,
  3 jitter strategies: uniform/constant/linear, monotonic TSC invariant)
- Phase 5: Canary (guard-page memory scanner detector with VEH callback,
  CanaryHandshake struct for engine-target coordination)
- Phase 6: AllocTracker enhancements (TrackedPage extended with modified/wasExecutable/
  allocAsRW/reencryptCycle, re-encrypt detection on RW->RX->RW transitions,
  full register emulation in HandleGuardPage via __cpuidex/__rdtsc,
  dirty skip in RescanTrackedPages)
- Phase 7: handshake_test tool (tests all 15 magic CPUID leaves, brand string,
  hypervisor hiding/bit)
- All 6 code review bugs fixed (P0-P2)
- 0 warnings 0 errors at /W4/WX across all 17 targets
- All branding/credits removed
- Phase A stealth model fixed
- All docs updated for new features

+++ src/engine/emu/ThreadManager.cpp
@@ -35,9 +35,9 @@ bool ThreadManager::HandleNtCreateThread(uint64_t* args, uint64_t* result)
 
     if (hThread) {
         if (args[0]) *(HANDLE*)(uintptr_t)args[0] = hThread;
-        *result = STATUS_SUCCESS;
+        *result = (uint64_t)STATUS_SUCCESS;
     } else {
-        *result = STATUS_ACCESS_DENIED;
+        *result = (uint64_t)STATUS_ACCESS_DENIED;
     }
 
     m_logger->Trace(LOG_EMU, "NtCreateThread -> hThread=0x%p", hThread);
